Moved Modifier Functions
- removed modifier functions from algorithms.py
- moved to object_ops.py
