+++ CHANGELOG.md
@@ -8,20 +8,21 @@ All changes will be documented here
 
 - Hair System added
 - Human Rotation Limits added
+- New option to fit in reverse direction to convert a character-specific mesh to a generic proxy.
+- New options to do more precise fitting by turning off some normally useful smoothing passes.
 
 ## Changed
 
 - Afro Female now the default character on startup
-- New option to fit in reverse direction to convert a character-specific mesh to a generic proxy.
-- Fixed fitting of meshes that are very close to the skin.
-- New options to do more precise fitting by turning off some normally useful smoothing passes.
 - Updated GUI
+- Removed Docs from project directory to it's own repository
 
 ## Bug Fixes
 
 - Certain bones have to be precisely aligned to the rotation of a specific other bone, which required new code.
 - Stretch to has to be applied to the muscle bones while fitting.
 - B-Bone handles and certain constraints have to be fixed in the library file via a script.
+- Fixed fitting of meshes that are very close to the skin.
 - Replaced obj.hide_select = False with obj.hide_set(False) fixing Age update bug.
 - Minor GUI edits to fix drawing issues.
 - Fixed Bump and Displacement issues in MBLab_Skin2
@@ -29,7 +30,7 @@ All changes will be documented here
 
 ## Known Issues
 
-
+- THe Hair Engine does NOT apply a shader at this time
 
 # MB-Lab 1.7.6
