+++ humanoid.py
@@ -32,7 +32,7 @@ import json
 import operator
 
 
-from . import morphengine, skeletonengine, algorithms, proxyengine, materialengine, utils, file_ops
+from . import morphengine, skeletonengine, algorithms, proxyengine, materialengine, utils, file_ops, object_ops
 
 logger = logging.getLogger(__name__)
 
@@ -260,7 +260,7 @@ class Humanoid:
     def add_subdivision_modifier(self):
         obj = self.get_object()
         parameters = {"levels": 1, "render_levels": 2, "show_viewport": True, "show_in_editmode": False}
-        algorithms.new_modifier(obj, self.mat_engine.subdivision_modifier_name, 'SUBSURF', parameters)
+        object_ops.new_modifier(obj, self.mat_engine.subdivision_modifier_name, 'SUBSURF', parameters)
 
     def add_displacement_modifier(self):
         obj = self.get_object()
@@ -268,7 +268,7 @@ class Humanoid:
         if disp_img:
             disp_tex = file_ops.new_texture(self.mat_engine.generated_disp_modifier_ID, disp_img)
             parameters = {"texture_coords":'UV', "strength": 0.01, "show_viewport": False, "texture": disp_tex}
-            displacement_modifier = algorithms.new_modifier(obj, self.mat_engine.generated_disp_modifier_ID, 'DISPLACE', parameters)
+            displacement_modifier = object_ops.new_modifier(obj, self.mat_engine.generated_disp_modifier_ID, 'DISPLACE', parameters)
 
     def rename_obj(self, prefix):
         obj = self.get_object()
@@ -473,38 +473,38 @@ class Humanoid:
 
     def get_subd_visibility(self):
         obj = self.get_object()
-        modfr = algorithms.get_modifier(obj, self.mat_engine.subdivision_modifier_name)
-        return algorithms.get_modifier_viewport(modfr)
+        modfr = object_ops.get_modifier(obj, self.mat_engine.subdivision_modifier_name)
+        return object_ops.get_modifier_viewport(modfr)
 
 
     def set_subd_visibility(self, value):
         obj = self.get_object()
-        modfr = algorithms.get_modifier(obj, self.mat_engine.subdivision_modifier_name)
-        algorithms.set_modifier_viewport(modfr, value)
+        modfr = object_ops.get_modifier(obj, self.mat_engine.subdivision_modifier_name)
+        object_ops.set_modifier_viewport(modfr, value)
 
 
     def set_smooth_visibility(self, value):
         obj = self.get_object()
-        modfr = algorithms.get_modifier(obj, self.corrective_modifier_name)
-        algorithms.set_modifier_viewport(modfr, value)
+        modfr = object_ops.get_modifier(obj, self.corrective_modifier_name)
+        object_ops.set_modifier_viewport(modfr, value)
 
 
     def get_smooth_visibility(self):
         obj = self.get_object()
-        modfr = algorithms.get_modifier(obj, self.corrective_modifier_name)
-        return algorithms.get_modifier_viewport(modfr)
+        modfr = object_ops.get_modifier(obj, self.corrective_modifier_name)
+        return object_ops.get_modifier_viewport(modfr)
 
     def get_disp_visibility(self):
 
         obj = self.get_object()
-        modfr = algorithms.get_modifier(obj, self.mat_engine.generated_disp_modifier_ID)
-        return algorithms.get_modifier_viewport(modfr)
+        modfr = object_ops.get_modifier(obj, self.mat_engine.generated_disp_modifier_ID)
+        return object_ops.get_modifier_viewport(modfr)
 
 
     def set_disp_visibility(self, value):
         obj = self.get_object()
-        modfr = algorithms.get_modifier(obj, self.mat_engine.generated_disp_modifier_ID)
-        algorithms.set_modifier_viewport(modfr, value)
+        modfr = object_ops.get_modifier(obj, self.mat_engine.generated_disp_modifier_ID)
+        object_ops.set_modifier_viewport(modfr, value)
 
 
     def sync_obj_props_to_character_materials(self):
@@ -848,7 +848,7 @@ class Humanoid:
                     props_to_process.remove(prop)
 
         for prop in props_to_process:
-            new_val = file_ops.generate_parameter(
+            new_val = algorithms.generate_parameter(
                 self.character_data[prop],
                 random_value,
                 prv_phenotype)
@@ -1198,4 +1198,4 @@ class Humanoid:
     def add_corrective_smooth_modifier(self):
         obj = self.get_object()
         parameters = {"show_viewport": True, "invert_vertex_group": True, "vertex_group": "head"}
-        algorithms.new_modifier(obj, self.corrective_modifier_name, 'CORRECTIVE_SMOOTH', parameters)
+        object_ops.new_modifier(obj, self.corrective_modifier_name, 'CORRECTIVE_SMOOTH', parameters)
