+++ material_ops.py
@@ -29,14 +29,17 @@ import bpy
 
 logger = logging.getLogger(__name__)
 
-#TODO Send to material_ops
+# ------------------------------------------------------------------------
+#    Material Functions
+# ------------------------------------------------------------------------
+
 def set_node_image(mat_node, mat_image):
     if mat_node:
         mat_node.image = mat_image
     else:
         logger.warning("Node assignment failed. Image not found: %s", mat_image)
 
-#
+
 def get_material(material_name):
     for material in bpy.data.materials:
         if material.name == material_name:
@@ -44,7 +47,7 @@ def get_material(material_name):
     logger.warning("Material %s not found", material_name)
     return None
 
-#
+
 def get_material_nodes(material):
     if material.node_tree:
         return material.node_tree.nodes
@@ -52,7 +55,7 @@ def get_material_nodes(material):
     logger.warning("Material %s has not nodes", material.name)
     return None
 
-#
+
 def get_material_node(material_name, node_name):
     material_node = None
     material = get_material(material_name)
@@ -63,7 +66,7 @@ def get_material_node(material_name, node_name):
         logger.warning("Node not found: %s in material %s", node_name, material_name)
     return material_node
 
-#
+
 def get_node_output_value(node, index):
     if index < len(node.outputs):
         if hasattr(node.outputs[index], "default_value"):
@@ -75,7 +78,7 @@ def get_node_output_value(node, index):
     logger.warning("Socket [%s] not in range of node %s outputs", index, node.name)
     return None
 
-#
+
 def set_node_output_value(node, index, value):
     if index < len(node.outputs):
         if hasattr(node.outputs[index], "default_value"):
